+++ lib/telegram_github_changes_bot.rb
@@ -2,3 +2,4 @@
 
 require 'telegram/bot'
 require_relative 'telegram_github_changes_bot/github_repo_changes'
+require_relative 'telegram_github_changes_bot/telegram_helper'