+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Fixes
 
 * Fix warning for `bundle install` (user `bundle config set without`)
+* Sanitize output to telegram (repleace `_`)
 
 ## 0.0.2